refactor(cmd): move dump.go to its own cmd/cluster/dump package
dump.go imports hypershift-operator/controllers/sharedingress, which
product-cli inherits transitively through cmd/cluster/core even though
it never exposes the dump command. Isolating the dump code in its own
package breaks this transitive dependency path.
