+++ cmd/cluster/dump/dump_test.go
@@ -1,4 +1,4 @@
-package core
+package dump
 
 import (
 	"context"
